Test Driven Development (TDD) .NET MVC
Automated testing can help ensure that your project behaviour falls
within expected boundaries.  Including testing in extremely large
projects can help eliminate vulnerabilities with automated checks for
expected responses under specific conditions.
Testing is sometimes encouraged during development – this practice is
called Test Driven Development (TDD).  TDD is useful since relevant
tests can be added while the developer is focused on the module.  The
danger of TDD is too much time can be spent writing tests on code that
is eventually replaced or thrown out.
Regardless of whether or not automated testing is useful for your
project, it is important to be aware of how to test so you can make a
better decision as to when and how you wish to test your code.